+++ itambox/assets/forms/filter_forms.py
@@ -11,6 +11,7 @@ from ..filters import (
     CategoryFilterSet,
     DepreciationFilterSet,
     ManufacturerFilterSet,
+    RepairEpisodeFilterSet,
     StatusLabelFilterSet,
     SupplierFilterSet,
     WarrantyFilterSet,
@@ -121,3 +122,7 @@ class WarrantyFilterForm(FilterForm):
 
 class AssetReservationFilterForm(FilterForm):
     filterset_class = AssetReservationFilterSet
+
+
+class RepairEpisodeFilterForm(FilterForm):
+    filterset_class = RepairEpisodeFilterSet